requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.8
django-allauth==0.34.0
django-daterange-filter==1.3.0
django-filter==1.1.0
django-jet==1.0.7
django-map-widgets==0.1.9
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
djangorestframework==3.7.7
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
oauth2client==4.1.2
oauthlib==2.0.6
openapi-codec==1.3.2
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
